# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Solo D&D 5e Companion App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns/{campaignId}/messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: List all chat messages for a campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
  /campaigns/{campaignId}/chat:
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message to the AI Dungeon Master (streaming SSE)
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "200":
          description: SSE stream of AI response
          content:
            text/event-stream:
              schema:
                type: string
...
  /campaigns/{campaignId}/inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: Get inventory for a campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: addInventoryItem
      tags: [inventory]
      summary: Add item to inventory
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemBody"
      responses:
        "201":
          description: Added item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /campaigns/{campaignId}/inventory/{itemId}:
    put:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update an inventory item
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemBody"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Remove item from inventory
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        diceRolls:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/RollResult"
        createdAt:
          type: string
          format: date-time
      required: [id, campaignId, role, content, createdAt]
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
        diceRolls:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/RollResult"
      required: [content]
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        quantity:
          type: integer
        weight:
          type: number
          nullable: true
        isEquipped:
          type: boolean
        itemType:
          type: string
          enum: [weapon, armor, consumable, tool, treasure, misc]
        createdAt:
          type: string
          format: date-time
      required: [id, campaignId, name, quantity, isEquipped, itemType, createdAt]
    CreateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        quantity:
          type: integer
        weight:
          type: number
          nullable: true
        isEquipped:
          type: boolean
        itemType:
          type: string
          enum: [weapon, armor, consumable, tool, treasure, misc]
      required: [name, quantity, isEquipped, itemType]
    UpdateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        quantity:
          type: integer
        weight:
          type: number
          nullable: true
        isEquipped:
          type: boolean
        itemType:
          type: string
          enum: [weapon, armor, consumable, tool, treasure, misc]
...
    RollResult:
      type: object
      properties:
        expression:
          type: string
        label:
          type: string
          nullable: true
        total:
          type: integer
        rolls:
          type: array
          items:
            type: integer
        modifier:
          type: integer
        details:
          type: string
      required: [expression, total, rolls, modifier, details]